Use int16 for change map raster

+++ algorithms/models/area_analysis/manager.py
@@ -110,7 +110,7 @@ class AreaAnalyst(QObject):
 
     def makeChangeMap(self):
         rows, cols = self.geodata['ySize'], self.geodata['xSize']
-        band = np.zeros([rows, cols])
+        band = np.zeros([rows, cols], dtype=np.int16)
 
         f, s = self.first, self.second
         if self.initRaster == None:
@@ -132,7 +132,7 @@ class AreaAnalyst(QObject):
                         else:
                             band[i, j] = self.encode(r, c)
                 self.updateProgress.emit()
-            bands = [np.ma.array(data = band, mask = f.mask)]
+            bands = [np.ma.array(data = band, mask = f.mask, dtype=np.int16)]
             raster = Raster()
             raster.create(bands, self.geodata)
             self.changeMap = raster
